+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Reset subscribers on code reloading to avoid double execution.
+
 ## 1.5.0 (2024-04-26)
 
 - Require Ruby 2.7+.